+++ src/tokens/functional/color/dark/presentational-dark.json5
@@ -1,6 +1,6 @@
 {
-  color: {
-    presentational: {
+  base: {
+    color: {
       gray: {
         '0': {
           $value: '#1c1c1c',
