+++ src/tokens/functional/color/light/presentational-light.json5
@@ -1,6 +1,6 @@
 {
-  color: {
-    presentational: {
+  base: {
+    color: {
       gray: {
         '0': {
           $value: '#e8ecf2',
